fix problem with context load inside tests
