+++ src/test/java/it/gov/pagopa/rtp/activator/RtpActivatorApplicationTest.java
@@ -1,16 +1,21 @@
 package it.gov.pagopa.rtp.activator;
 
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.ApplicationContext;
+
 
 @SpringBootTest
 class RtpActivatorApplicationTests {
 
+  @Autowired
+  private ApplicationContext applicationContext;
+
   @Test
   void contextLoads() {
-    assertTrue(true);
+    assertNotNull(applicationContext);
   }
-
 }
